requirements.txt
asgiref==3.6.0
Django==4.0
djangorestframework==3.13.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
